Fill a stack entry from all its commits and gate on unopened branches
- fill the form from several commits as Submit stack describes an
  entry: the first subject as title, the rest listed in the body
- keep entries above a branch with no open pull request out of
  reach in the pull request tab, pushed or not, since a pull request
  cannot target a base GitHub has none for
- copy any branch name in the stack strip from its menu
